Stats_Bot/requirements.txt
discord.py==2.3.1
beautifulsoup4==4.12.2
pytz==2023.3